Issue #67 - access_token is stored in global conf not local .aio
This change checks where an IMS context is defined,
and stores the tokens in the same location (i.e. local
or global).
Furthermore, on logout, _only_ the tokens are removed,
rather than the entire IMS context being rewritten,
as was the case previously. The context data is the
merged view of global and local configs for a context,
so rewriting it risks accidentally copying data between
the local IMS context definition and the global one.